Revert "Feat: BoxedExpression.constantValue ('getter' property)"
This reverts commit a8f85fa3328503d31825562e3ed573b74f73c30f.

This property/getter is largely unnecessary: because 'value' largely is  'constantValue' (the exception is for symbols: in which for non-constants its value may vary arbitrarily, or with assumptions).

Note:
- This *could* be introduced, perhaps exclusively for symbols; or its overall definition could be adjusted slightly such that it is useful for boxed-functions too.
I.e., for a boxed-function, only return a value if 'isConstant' && 'isPure'. ('value' differs in that returns non-undefined if 'isPure', *only*...)